Fix player rotation issues for high-latency clients
The player motion is split into two phases, one tick apart.

In the first tick the visible motion is update with the movment
expected during one tick (0.2 for slow speed and 1.0 while
holding the Book of Dragons), but the horse and player are not
moved. This allows one full tick for the client to server update
packet to arrive and be processed.

In the second tick both the visible elements and the player are
moved two ticks worth of distance, i.e. 0.4 / 2.0.

This will introduce visible artifacts since the player is only updated
every second tick, making the dragon appear to move forward/back a
little every other tick from the player point of view.

Background:
It seems that teleporting the horse the player rides temporarily
resets the Rotation attribute of the playter, server side. However,
the client side is in control for player mounted hourses and thus
this erronous rotation is corrected next tick. For long latency
client connections this correction may arrive too late so we end
up in a loop of perpetually server-overridden player Rotation
values until chance makes the server late, giving the client
packet a chance to update the server state.