allow the mount point to glow for a short while when summoned
1.21.6 and possibly earlier exhibit a client side bug where the dragon
rotation is not shown correct after sumoning, making it difficult to
mount the dragon. This is corrected when logging out/in again or after
mounting the dragon and starting to fly it.

As a workaround the mount can be made to glow for a short period of time
after summoning a dragon. This makes it easier to click the correct
spot.